# resample: Port keyframe resampling to WebAssembly