--- a/packages/functions/package.json
+++ b/packages/functions/package.json
@@ -37,7 +37,7 @@
 	"dependencies": {
 		"@gltf-transform/core": "^3.2.1",
 		"@gltf-transform/extensions": "^3.2.1",
-		"keyframe-resample": "^0.0.7",
+		"keyframe-resample": "^0.0.8",
 		"ktx-parse": "^0.5.0",
 		"ndarray": "^1.0.19",
 		"ndarray-lanczos": "^0.3.0",