As discussed in #922, the resample() function currently spends a lot of time in resource graph operations. Rewriting it to operate directly on the underlying accessor arrays improves performance dramatically (170ms → 60ms for my 800,000 keyframe test dataset on M1 mac). Rewriting the logic in optimized WebAssembly gets even greater savings (170ms → 20ms).

The WebAssembly code is compiled from AssemblyScript, and will be maintained in a separate repository (https://github.com/donmccurdy/keyframe-resample-wasm).  

***

- Fixes #922 
- Fixes #290 

***

Major thanks to @kzhsw for extensive research, prototyping, and advice in #922.