@kzhsw 2023-05-02
Things to consider:
* [keyframe-resample-wasm](https://github.com/donmccurdy/keyframe-resample-wasm) uses a [shared](https://github.com/donmccurdy/keyframe-resample-wasm/blob/d6896fa73af70370abf711486cdcc9c0b9b861ba/src/resample_bindings.ts#L32) wasm instance, maybe users can be given options to release it to release memory.
* Inline the wasm in js like [meshoptimizer](https://github.com/zeux/meshoptimizer) for old bundlers
* Multithreading for large chunks or different channels
* Fallback to js impl on unsupported plarform of user's choice