@donmccurdy 2023-05-02
There are several other WASM dependencies in the project now (Draco, Meshopt, MikkTSpace), so this module isn't necessarily the limiting factor on bundler and platform support.

I think I'll need to draw a line somewhere on the complexity, or this isn't going to get finished.  😇 If you (or another reader) needs one of those things, I'd be happy to discuss tradeoffs around it. But unless there's a clear need, I'd like to keep this from getting any more complex than it is now.