@kzhsw 2023-05-02
> There are several other WASM dependencies in the project now (Draco, Meshopt, MikkTSpace), so this module isn't necessarily the limiting factor on bundler and platform support.

The existing WASM dependencies are used by passing the wasm impl from options, that could make a difference for bundlers.

Also, should this be consdered as breaking change since this might require changes of some bundler users?