@donmccurdy 2023-05-05
Here's where I ended up on backward-compatibility:

- The /functions module inlines the *JavaScript* implementation from the `keyframe-resample` module, which is still slow, but faster than the original implementation and should be fully backward compatible. No WASM binary is loaded by default.
- Users of the /functions module may optionally provide a WASM implementation from `keyframe-resample` or `keyframe-resample/compat` themselves. The latter is compatible with older bundlers, using a Data URI for the WASM build.
- The /cli module automatically uses the faster WASM implementation.